refactor(spf): share Mux identity across the SPF Medias
`src`, `source`, and `contentData` are Mux identity, not engine behavior, but
they lived on the class built over the full HLS engine. The audio-only Mux Media
needs the same three over a different SPF Media, and since each extends a
different base there is no common class to put them on — only a common `src`
accessor to write through, which is what a mixin wants.

`MuxMediaMixin` lands in its own module deliberately: it imports
`@videojs/media/dom/mux/source` and nothing else, so the audio-only adapter can
take it without pulling the full HLS engine along. Importing from `../mux` or
`../mux/media` instead would, since those reach `SimpleHlsMedia`.

Behavior-preserving — the existing 16 tests pass untouched, and `./hls` stays at
19.36 KB gzipped.
